Instantiate aliases
In some cases, the using aliases points
to a template instantiated class, e.g. A<3>
An alias does NOT trigger the template instantiation in clang,
(it is lazy).
Previously we reported an error, asking the user to explicitly
instantiate the class.
New behaviour: I call the Sema to make the explicit instantiation
on the fly (with proper diagnostics) and proceed.